+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.17 wip
 * repl some reflection warnings removed
+* multi method `exists?` added to `tools.io.core` with its proxy in `tools.io`
 
 ## 0.3.16 (2019/02/22)